+++ api/src/main/java/com/dnd/sbooky/api/docs/spec/DrawItemApiSpec.java
@@ -1,6 +1,6 @@
 package com.dnd.sbooky.api.docs.spec;
 
-import com.dnd.sbooky.api.item.response.DrawItemResponse;
+import com.dnd.sbooky.api.item.v1.response.DrawItemResponse;
 import com.dnd.sbooky.api.support.response.ApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
