+++ api/src/main/java/com/dnd/sbooky/api/docs/spec/FindItemsApiSpec.java
@@ -1,7 +1,7 @@
 package com.dnd.sbooky.api.docs.spec;
 
-import com.dnd.sbooky.api.item.response.FindItemsResponse;
-import com.dnd.sbooky.api.item.response.FindMemberEquippedItemsResponse;
+import com.dnd.sbooky.api.item.v1.response.FindItemsResponse;
+import com.dnd.sbooky.api.item.v1.response.FindMemberEquippedItemsResponse;
 import com.dnd.sbooky.api.support.response.ApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
