+++ api/src/main/java/com/dnd/sbooky/api/item/v1/DrawItemUsecase.java
@@ -1,9 +1,9 @@
-package com.dnd.sbooky.api.item;
+package com.dnd.sbooky.api.item.v1;
 
 import static com.dnd.sbooky.api.support.error.ErrorType.*;
 
-import com.dnd.sbooky.api.item.exception.ItemNotFoundException;
-import com.dnd.sbooky.api.item.response.DrawItemResponse;
+import com.dnd.sbooky.api.item.v1.exception.ItemNotFoundException;
+import com.dnd.sbooky.api.item.v1.response.DrawItemResponse;
 import com.dnd.sbooky.api.member.exception.MemberNotFoundException;
 import com.dnd.sbooky.api.point.AccumulatePointUseCase;
 import com.dnd.sbooky.core.item.ItemCode;
