+++ api/src/main/java/com/dnd/sbooky/api/item/v1/FindEquippedItemUsecase.java
@@ -1,6 +1,6 @@
-package com.dnd.sbooky.api.item;
+package com.dnd.sbooky.api.item.v1;
 
-import com.dnd.sbooky.api.item.response.FindEquippedItemsResponse;
+import com.dnd.sbooky.api.item.v1.response.FindEquippedItemsResponse;
 import com.dnd.sbooky.core.item.ItemCode;
 import com.dnd.sbooky.core.item.ItemType;
 import com.dnd.sbooky.core.item.MemberItemRepository;
