+++ api/src/main/java/com/dnd/sbooky/api/item/v1/SwitchEquippedItemController.java
@@ -1,7 +1,7 @@
-package com.dnd.sbooky.api.item;
+package com.dnd.sbooky.api.item.v1;
 
 import com.dnd.sbooky.api.docs.spec.SwitchEquippedItemApiSpec;
-import com.dnd.sbooky.api.item.request.SwitchEquippedItemRequest;
+import com.dnd.sbooky.api.item.v1.request.SwitchEquippedItemRequest;
 import com.dnd.sbooky.api.support.response.ApiResponse;
 import io.swagger.v3.oas.annotations.Parameter;
 import jakarta.validation.Valid;
