+++ api/src/main/java/com/dnd/sbooky/api/item/v1/exception/MemberHasNotItemException.java
@@ -1,4 +1,4 @@
-package com.dnd.sbooky.api.item.exception;
+package com.dnd.sbooky.api.item.v1.exception;
 
 import com.dnd.sbooky.api.support.error.ApiException;
 import com.dnd.sbooky.api.support.error.ErrorType;
